# docs/INF/openapi/teralya-openapi-v1.0.yaml
openapi: 3.1.0
...
info:
  title: Teralya API
  version: 1.0.0
  description: Contrato OpenAPI 3.1 del MVP. Derivado de INF-08 v2.5, INF-10 v1.0 e INF-10-A v1.0.
  contact:
    name: Teralya
    url: https://teralya.es
  license:
    name: Propietario — Teralya
    identifier: LicenseRef-Teralya-Proprietary
servers:
- url: https://api.teralya.es
  description: Producción
...
paths:
...
  /bodegas:
    post:
      operationId: api005
      x-teralya-api-code: API-005
      x-teralya-cu:
      - CU-012
      x-teralya-pt:
      - PT-ACC-002
      tags:
      - Bodegas
      summary: Solicitar registro de bodega
      description: API-005 registra una solicitud enviada con estado pendiente_revision y tipo estandar. La condición fundadora no se concede automáticamente (DLOG 0022).
      security: []
      responses:
        '201':
          description: Operación completada
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BodegaSelf'
              examples:
                success:
                  value:
                    id: 11111111-1111-4111-8111-111111111111
                    nombre_comercial: aaaaa
                    estado: pendiente_revision
                    created_at: '2026-07-13T12:00:00Z'
                    updated_at: '2026-07-13T12:00:00Z'
        '400':
          description: Validation Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/validation-error
                    title: Validation Error
                    status: 400
                    detail: La solicitud no pudo completarse.
                    code: VALIDATION_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '409':
          description: Conflict
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/conflict
                    title: Conflict
                    status: 409
                    detail: La solicitud no pudo completarse.
                    code: CONFLICT
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '500':
          description: Internal Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/internal-error
                    title: Internal Error
                    status: 500
                    detail: La solicitud no pudo completarse.
                    code: INTERNAL_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BodegaRegistrationRequest'
            examples:
              success:
                value:
                  nombre_comercial: aaaaa
                  razon_social: aaaaa
                  cif_vat: aaaaa
                  email: usuario@example.com
                  password: aaaaaaaaaaaa
                  persona_contacto: aaaaa
                  telefono: aaaaa
                  aceptacion_condiciones: true
  /bodegas/yo/perfil:
    patch:
      operationId: api006
      x-teralya-api-code: API-006
      x-teralya-cu:
      - CU-014
      x-teralya-pt:
      - PT-BOD-002
      tags:
      - Bodegas
      summary: Actualizar perfil propio
      description: Contrato normativo de API-006.
      security:
      - BearerOpaque: []
      responses:
        '200':
          description: Operación completada
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BodegaSelf'
              examples:
                success:
                  value:
                    id: 11111111-1111-4111-8111-111111111111
                    nombre_comercial: aaaaa
                    estado: borrador
                    created_at: '2026-07-13T12:00:00Z'
                    updated_at: '2026-07-13T12:00:00Z'
        '400':
          description: Validation Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/validation-error
                    title: Validation Error
                    status: 400
                    detail: La solicitud no pudo completarse.
                    code: VALIDATION_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '401':
          description: Authentication Required
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            WWW-Authenticate:
              description: Desafío Bearer según RFC 6750
              schema:
                type: string
                const: Bearer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/authentication-required
                    title: Authentication Required
                    status: 401
                    detail: La solicitud no pudo completarse.
                    code: AUTHENTICATION_REQUIRED
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '403':
          description: Forbidden
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/forbidden
                    title: Forbidden
                    status: 403
                    detail: La solicitud no pudo completarse.
                    code: FORBIDDEN
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '409':
          description: Conflict
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/conflict
                    title: Conflict
                    status: 409
                    detail: La solicitud no pudo completarse.
                    code: CONFLICT
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '500':
          description: Internal Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/internal-error
                    title: Internal Error
                    status: 500
                    detail: La solicitud no pudo completarse.
                    code: INTERNAL_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BodegaProfilePatch'
            examples:
              success:
                value:
                  nombre_comercial: aaaaa
    get:
      operationId: api031
      x-teralya-api-code: API-031
      x-teralya-cu:
      - CU-014
      x-teralya-pt:
      - PT-BOD-002
      tags:
      - Bodegas
      summary: Consultar perfil propio
      description: Contrato normativo de API-031.
      security:
      - BearerOpaque: []
      responses:
        '200':
          description: Operación completada
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BodegaSelf'
              examples:
                success:
                  value:
                    id: 11111111-1111-4111-8111-111111111111
                    nombre_comercial: aaaaa
                    estado: borrador
                    created_at: '2026-07-13T12:00:00Z'
                    updated_at: '2026-07-13T12:00:00Z'
        '401':
          description: Authentication Required
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            WWW-Authenticate:
              description: Desafío Bearer según RFC 6750
              schema:
                type: string
                const: Bearer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/authentication-required
                    title: Authentication Required
                    status: 401
                    detail: La solicitud no pudo completarse.
                    code: AUTHENTICATION_REQUIRED
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '403':
          description: Forbidden
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/forbidden
                    title: Forbidden
                    status: 403
                    detail: La solicitud no pudo completarse.
                    code: FORBIDDEN
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '404':
          description: Resource Not Found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/resource-not-found
                    title: Resource Not Found
                    status: 404
                    detail: La solicitud no pudo completarse.
                    code: RESOURCE_NOT_FOUND
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '500':
          description: Internal Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/internal-error
                    title: Internal Error
                    status: 500
                    detail: La solicitud no pudo completarse.
                    code: INTERNAL_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
...
  /bodegas/{id}:
    get:
      operationId: api030
      x-teralya-api-code: API-030
      x-teralya-cu:
      - CU-006
      x-teralya-pt:
      - PT-PUB-004
      tags:
      - Bodegas
      summary: Consultar bodega pública
      description: Contrato normativo de API-030.
      security: []
      responses:
        '200':
          description: Operación completada
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BodegaPublic'
              examples:
                success:
                  value:
                    id: 11111111-1111-4111-8111-111111111111
                    nombre_comercial: aaaaa
                    vinos: []
        '404':
          description: Resource Not Found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/resource-not-found
                    title: Resource Not Found
                    status: 404
                    detail: La solicitud no pudo completarse.
                    code: RESOURCE_NOT_FOUND
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '500':
          description: Internal Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/internal-error
                    title: Internal Error
                    status: 500
                    detail: La solicitud no pudo completarse.
                    code: INTERNAL_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
        description: Identificador id
        example: 11111111-1111-4111-8111-111111111111
...
  /admin/bodegas/{id}:
    get:
      operationId: api036
      x-teralya-api-code: API-036
      x-teralya-cu:
      - CU-021
      x-teralya-pt:
      - PT-ADM-003
      tags:
      - Administración
      summary: Consultar bodega administrativa
      description: Contrato normativo de API-036.
      security:
      - BearerOpaque: []
      responses:
        '200':
          description: Operación completada
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BodegaAdmin'
              examples:
                success:
                  value:
                    id: 11111111-1111-4111-8111-111111111111
                    nombre_comercial: aaaaa
                    estado: borrador
                    created_at: '2026-07-13T12:00:00Z'
                    updated_at: '2026-07-13T12:00:00Z'
        '401':
          description: Authentication Required
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
            WWW-Authenticate:
              description: Desafío Bearer según RFC 6750
              schema:
                type: string
                const: Bearer
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/authentication-required
                    title: Authentication Required
                    status: 401
                    detail: La solicitud no pudo completarse.
                    code: AUTHENTICATION_REQUIRED
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '403':
          description: Forbidden
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/forbidden
                    title: Forbidden
                    status: 403
                    detail: La solicitud no pudo completarse.
                    code: FORBIDDEN
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '404':
          description: Resource Not Found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/resource-not-found
                    title: Resource Not Found
                    status: 404
                    detail: La solicitud no pudo completarse.
                    code: RESOURCE_NOT_FOUND
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
        '500':
          description: Internal Error
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
              examples:
                error:
                  value:
                    type: https://teralya.es/problems/internal-error
                    title: Internal Error
                    status: 500
                    detail: La solicitud no pudo completarse.
                    code: INTERNAL_ERROR
                    request_id: 11111111-1111-4111-8111-111111111111
                    retryable: false
      parameters:
      - name: id
        in: path
        required: true
        schema: *id001
        description: Identificador id
        example: 11111111-1111-4111-8111-111111111111
...
components:
  securitySchemes:
    BearerOpaque:
      type: http
      scheme: bearer
      description: Token opaco de 256 bits; TTL absoluto 8 horas.
  headers:
    XRequestId:
      description: Identificador UUID de trazabilidad
      schema: *id001
  schemas:
    FieldError:
      type: object
      additionalProperties: false
      properties:
        field: &id013
          type: string
          minLength: 1
          maxLength: 200
        code: *id006
        message: &id007
          type: string
          minLength: 1
          maxLength: 500
      required:
      - field
      - code
      - message
    Problem:
      type: object
      additionalProperties: false
      properties:
        type: &id016
          type: string
          format: uri
          minLength: 1
          maxLength: 2048
        title: *id002
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail: *id007
        instance:
          type: string
          minLength: 1
          maxLength: 2048
        code: *id006
        request_id: *id001
        field_errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
          maxItems: 50
        retryable:
          type: boolean
      required:
      - type
      - title
      - status
      - detail
      - code
      - request_id
      - retryable
...
    BodegaRegistrationRequest:
      type: object
      additionalProperties: false
      properties:
        nombre_comercial: *id002
        razon_social: *id013
        cif_vat: &id014
          type: string
          minLength: 1
          maxLength: 32
        email: *id010
        password: *id011
        persona_contacto: *id009
        telefono: *id014
        aceptacion_condiciones:
          type: boolean
          const: true
        pais_contacto: *id009
        ciudad: *id009
        codigo_postal:
          type: string
          minLength: 1
          maxLength: 20
      required:
      - nombre_comercial
      - razon_social
      - cif_vat
      - email
      - password
      - persona_contacto
      - telefono
      - aceptacion_condiciones
    BodegaProfilePatch:
      type: object
      additionalProperties: false
      properties:
        nombre_comercial: *id002
        historia: &id015
          type: string
          minLength: 1
          maxLength: 5000
        filosofia: *id015
        region: *id002
        pais: *id009
        denominacion_origen: *id002
        anio_fundacion: &id017
          type: integer
          minimum: 1800
          maximum: 2100
        web: *id016
        video_url: *id016
        email_principal: *id010
        telefono: *id014
        persona_contacto: *id009
        logo_url: *id016
        imagen_principal_url: *id016
      minProperties: 1
    BodegaSummary:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        nombre_comercial: *id002
        slug: *id002
        logo_url: *id016
        region: *id002
        pais: *id009
        denominacion_origen: *id002
      required:
      - id
      - nombre_comercial
    BodegaPublic:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        nombre_comercial: *id002
        vinos:
          type: array
          items:
            $ref: '#/components/schemas/WineSummary'
        slug: *id002
        logo_url: *id016
        imagen_principal_url: *id016
        historia: *id015
        filosofia: *id015
        region: *id002
        pais: *id009
        denominacion_origen: *id002
        anio_fundacion: *id017
        web: *id016
        video_url: *id016
      required:
      - id
      - nombre_comercial
      - vinos
    BodegaSelf:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        nombre_comercial: *id002
        slug: *id002
        logo_url: *id016
        imagen_principal_url: *id016
        historia: *id015
        filosofia: *id015
        region: *id002
        pais: *id009
        denominacion_origen: *id002
        anio_fundacion: *id017
        web: *id016
        video_url: *id016
        estado: &id019
          type: string
          enum:
          - borrador
          - pendiente_revision
          - aprobada
          - activa
          - suspendida
          - archivada
        created_at: *id018
        updated_at: *id018
        razon_social: *id013
        cif_vat: *id014
        email_principal: *id010
        telefono: *id014
        persona_contacto: *id009
        direccion_fisica: *id013
        codigo_postal:
          type: string
          minLength: 1
          maxLength: 20
        ciudad: *id009
        provincia: *id009
        pais_contacto: *id009
      required:
      - id
      - nombre_comercial
      - estado
      - created_at
      - updated_at
...
    BodegaAdmin:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        nombre_comercial:
          type: string
          minLength: 1
          maxLength: 160
        slug:
          type: string
          minLength: 1
          maxLength: 160
        logo_url:
          type: string
          format: uri
          minLength: 1
          maxLength: 2048
        imagen_principal_url:
          type: string
          format: uri
          minLength: 1
          maxLength: 2048
        historia:
          type: string
          minLength: 1
          maxLength: 5000
        filosofia:
          type: string
          minLength: 1
          maxLength: 5000
        region:
          type: string
          minLength: 1
          maxLength: 160
        pais:
          type: string
          minLength: 1
          maxLength: 100
        denominacion_origen:
          type: string
          minLength: 1
          maxLength: 160
        anio_fundacion:
          type: integer
          minimum: 1800
          maximum: 2100
        web:
          type: string
          format: uri
          minLength: 1
          maxLength: 2048
        video_url:
          type: string
          format: uri
          minLength: 1
          maxLength: 2048
        estado:
          type: string
          enum:
          - borrador
          - pendiente_revision
          - aprobada
          - activa
          - suspendida
          - archivada
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        razon_social:
          type: string
          minLength: 1
          maxLength: 200
        cif_vat:
          type: string
          minLength: 1
          maxLength: 32
        email_principal:
          type: string
          format: email
          minLength: 3
          maxLength: 254
        telefono:
          type: string
          minLength: 1
          maxLength: 32
        persona_contacto:
          type: string
          minLength: 1
          maxLength: 100
        direccion_fisica:
          type: string
          minLength: 1
          maxLength: 200
        codigo_postal:
          type: string
          minLength: 1
          maxLength: 20
        ciudad:
          type: string
          minLength: 1
          maxLength: 100
        provincia:
          type: string
          minLength: 1
          maxLength: 100
        pais_contacto:
          type: string
          minLength: 1
          maxLength: 100
        fecha_alta: *id018
        fecha_aprobacion: *id018
      required:
      - id
      - nombre_comercial
      - estado
      - created_at
      - updated_at
...
    ImageSummary:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        url: *id016
        es_principal:
          type: boolean
        orden:
          type: integer
          minimum: 0
          maximum: 999
        alt_text: *id007
        resolucion: *id014
      required:
      - id
      - url
      - es_principal
      - orden
      - alt_text
    WineSummary:
      type: object
      additionalProperties: false
      properties:
        id: *id001
        nombre_comercial: *id002
        precio: *id003
        moneda: *id008
        disponible_venta:
          type: boolean
        bodega:
          $ref: '#/components/schemas/BodegaSummary'
        slug: *id002
        tipo_vino: *id006
        anada: *id017
        region: *id002
        denominacion_origen: *id002
        imagen_principal:
          $ref: '#/components/schemas/ImageSummary'
      required:
      - id
      - nombre_comercial
      - precio
      - moneda
      - disponible_venta
      - bodega
...